#200 [Fix] section 페이지 무한 루프 수정 (#201)
* 🐛fix: section api

* 🐛fix: 컴포넌트 작성

* Revert "🐛fix: 컴포넌트 작성"

This reverts commit a3c6afa5347fe6bcb0692d4848f2ffcaf6e395ba.

* Revert "🐛fix: section api"

This reverts commit 7eb64ce3be847c4142939b843f080af3b007ac36.

* 🐛fix: section api

* 🐛fix: 컴포넌트 작성

* 🐛fix: 정민님 피드백/.....

* 🐛fix: 자잘한 오류 수정
