+++ src/components/writeRetro/task/TeamTask.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import { BiLike, BiSolidLike } from 'react-icons/bi';
 import { CgProfile } from 'react-icons/cg';
 import { FaRegTrashAlt } from 'react-icons/fa';
@@ -16,10 +16,10 @@ import {
   PopoverTrigger,
   Portal,
 } from '@chakra-ui/react';
-import { formattedDate } from './PersonalTask';
 import TeamTaskMessage from './taskMessage/TeamTaskMessage';
 import { sectionData } from '@/api/@types/Section';
 import { SectionServices } from '@/api/services/Section';
+import { convertToLocalTime } from '@/components/RetroList/ContentsList';
 import ActionItemTask from '@/components/writeRetro/ActionItems/ActionItemTask';
 import ReviseModal from '@/components/writeRetro/task/ReviseModal';
 import { useCustomToast } from '@/hooks/useCustomToast';
@@ -27,16 +27,20 @@ import * as S from '@/styles/writeRetroStyles/Layout.style';
 
 interface Props {
   section: sectionData;
+  like: number;
+  content: string;
 }
 
-const TeamTask: FC<Props> = ({ section }) => {
+const TeamTask: FC<Props> = ({ section, like, content }) => {
+  const { search } = useLocation();
+  const query = search.split(/[=,&]/);
+  const retrospectiveId = Number(query[1]);
+  const teamId = Number(query[3]);
   const toast = useCustomToast();
   const [liked, setLiked] = useState<number>(0);
   const [messaged, setMessaged] = useState<boolean>(false);
   const [isVisible, setIsVisible] = useState<boolean>(false);
-  const { search } = useLocation();
 
-  const query = search.split(/[=,&]/);
   const rId = Number(query[1]); // action-items로 넘겨줄 Id값들
   const tId = Number(query[3]);
   const sId: number = section.sectionId;
@@ -45,7 +49,6 @@ const TeamTask: FC<Props> = ({ section }) => {
     try {
       const data = await SectionServices.likePost({ sectionId: section.sectionId });
       setLiked(data.data.likeCnt);
-      console.log('like', liked);
     } catch (e) {
       toast.error(e);
     }
@@ -64,6 +67,22 @@ const TeamTask: FC<Props> = ({ section }) => {
     }
   };
 
+  const fetchSection = async () => {
+    try {
+      if (!teamId) {
+        await SectionServices.PersonalGet({ retrospectiveId: retrospectiveId });
+      } else {
+        await SectionServices.TeamGet({ retrospectiveId: retrospectiveId, teamId: teamId });
+      }
+    } catch (e) {
+      toast.error(e);
+    }
+  };
+
+  useEffect(() => {
+    fetchSection();
+  }, [like, content]);
+
   return (
     <>
       <S.TaskBox>
@@ -72,7 +91,7 @@ const TeamTask: FC<Props> = ({ section }) => {
           <Flex margin="10px auto">
             <S.TaskUserProfile>
               <CgProfile size="40px" color="#DADEE5" />
-              <S.TaskUserName>{section.username}</S.TaskUserName>
+              <S.TaskUserName>{section.username ?? '닉네임 없음'}</S.TaskUserName>
             </S.TaskUserProfile>
 
             <Popover>
@@ -107,17 +126,17 @@ const TeamTask: FC<Props> = ({ section }) => {
               <S.TaskText>
                 {section.content}
                 {/* <S.ReviseText>(수정됨)</S.ReviseText> */}
-                <S.ManagerStyle>
-                  <div>
-                    <ActionItemTask tId={tId} rId={rId} sId={sId} />
-                  </div>
-                  <S.ManagerText>담당자</S.ManagerText>
-                </S.ManagerStyle>
+                {section.sectionName === 'Action Items' && (
+                  <S.ManagerStyle>
+                    <div>
+                      <ActionItemTask tId={tId} rId={rId} sId={sId} />
+                    </div>
+                    <S.ManagerText>담당자</S.ManagerText>
+                  </S.ManagerStyle>
+                )}
               </S.TaskText>
             </PopoverTrigger>
             <PopoverContent>
-              {/* TaskTextModal */}
-
               <ReviseModal section={section} />
             </PopoverContent>
           </Popover>
@@ -143,7 +162,7 @@ const TeamTask: FC<Props> = ({ section }) => {
               <S.SubTaskIcon>
                 <MdAccessAlarm size="20px" color="#DADEE5" />
               </S.SubTaskIcon>
-              <S.SubTaskCount>{formattedDate(section.createdDate)}</S.SubTaskCount>
+              <S.SubTaskCount>{convertToLocalTime(section.createdDate)}</S.SubTaskCount>
             </S.SubTaskStyle>
           </S.SubTaskBox>
         </S.TaskMainStyle>
